requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
psycopg[binary]==3.2.6
psycopg_pool==3.2.6
openai==1.75.0
httpx==0.28.1
apscheduler==3.10.4
pydantic-settings==2.8.1
